Upgraded Building code

+++ test code/Office Building Module/OfficeBuilding/lib/HAL/ISolarPanel.h
@@ -5,7 +5,7 @@ class ISolarPanel
 {  
     public:
     virtual ~ISolarPanel() = default;
-    virtual int getSolarPower() = 0;
+    virtual double getSolarPower() = 0;
 };
 
 #endif
